Some minor fixes fortrace info
