+++ packages/overlay/src/integrations/sentry/tabs/TracesTab.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
-import TraceDetails from '../components/traces/TraceDetails';
 import TraceList from '../components/traces/TraceList';
+import TraceDetails from '../components/traces/traceDetail';
 import { SentryEventsContextProvider } from '../data/sentryEventsContext';
 
 export default function TracesTab() {
